✨ feat: add city , changes mapper , dtos , exceptions and controllers

+++ caroneiros/src/main/java/caroneiros/services/CarpoolService.java
@@ -1,6 +1,7 @@
 package caroneiros.services;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,8 +12,10 @@ import caroneiros.domain.models.Vehicle;
 import caroneiros.domain.repositories.CarpoolRepository;
 import caroneiros.dtos.CarpoolRequestDTO;
 import caroneiros.dtos.CarpoolResponseDTO;
+import caroneiros.dtos.CarpoolUpdateRequestDTO;
 import caroneiros.dtos.mapper.CarpoolMapper;
 import caroneiros.infra.exceptions.DontDriverException;
+import caroneiros.infra.exceptions.InvalidOperationException;
 import caroneiros.infra.exceptions.NoVehiclesException;
 import caroneiros.infra.exceptions.NotFoundException;
 import caroneiros.infra.exceptions.VehicleNotOwnedException;
@@ -33,26 +36,24 @@ public class CarpoolService implements CarpoolServiceInterface {
     private VehicleService vehicleService;
 
     @Override
-    public CarpoolResponseDTO saveCarpool(CarpoolRequestDTO carpoolDTO) {
-       
-        AppUser driver = userService.findUserById(carpoolDTO.driverId());
+    public CarpoolResponseDTO saveCarpool(Long driverId, Long vehicleId, CarpoolRequestDTO carpoolDTO) {
+
+        AppUser driver = userService.findUserById(driverId);
 
         validateDriver(driver);
 
-        Vehicle vehicle = getVehicleOwnership(carpoolDTO, driver);
+        valideVehicleOwnership(vehicleId, driver);
 
         Carpool carpool = CarpoolMapper.toEntity(carpoolDTO, driver);
-
         this.saveCarpool(carpool);
-        return CarpoolMapper.toCarpoolResponseDTO(carpool, vehicle);
+        return CarpoolMapper.toCarpoolResponseDTO(carpool);
     }
 
-    private Vehicle getVehicleOwnership(CarpoolRequestDTO carpoolDTO, AppUser driver) {
-        Vehicle vehicle = vehicleService.getVehicleByIdOrThrow(carpoolDTO.vehicleId());
+    private void valideVehicleOwnership(Long vehicleId, AppUser driver) {
+        Vehicle vehicle = vehicleService.getVehicleByIdOrThrow(vehicleId);
         if (!driver.getVehicles().contains(vehicle)) {
             throw new VehicleNotOwnedException("O véiculo não pertence ao usuário informado");
         }
-        return vehicle;
     }
 
     private void validateDriver(AppUser driver) {
@@ -66,6 +67,10 @@ public class CarpoolService implements CarpoolServiceInterface {
     @Override
     public Carpool findCarpoolById(Long id) {
         log.info("Buscando por carona de id [{}]", id);
+        return getCarpoolByIdOrThrow(id);
+    }
+
+    private Carpool getCarpoolByIdOrThrow(Long id) {
         return carpoolRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Carona não encontrada!"));
     }
@@ -74,8 +79,9 @@ public class CarpoolService implements CarpoolServiceInterface {
     @Transactional
     public void deleteCarpoolById(Long id) {
         log.info("Deletando carona de id [{}]", id);
-        Carpool carpool = carpoolRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException("Carona não encontrada!"));
+        Carpool carpool = getCarpoolByIdOrThrow(id);
+        if (!carpool.getReservations().isEmpty())
+            throw new InvalidOperationException("Sua carona já possui reservas. Não é possivel deleta-la");
         carpoolRepository.delete(carpool);
 
     }
@@ -84,31 +90,41 @@ public class CarpoolService implements CarpoolServiceInterface {
     @Transactional
     public Carpool uptadeCarpool(Long id, Carpool carpoolToUpdate) {
         Carpool carpoolFromDB = findCarpoolById(id);
-
-        int reservedSeats = carpoolFromDB.getReservations().size();
-        if (carpoolToUpdate.getSeatsAvailable() < reservedSeats) {
-            throw new IllegalArgumentException(
-                    "A quantidade de assentos disponíveis não pode ser menor que o número de reservas confirmadas.");
-        }
         carpoolFromDB.setDriver(carpoolToUpdate.getDriver());
-        carpoolFromDB.setEstimatedDeparture(carpoolToUpdate.getEstimatedDeparture());
         carpoolFromDB.setEstimatedArrival(carpoolToUpdate.getEstimatedArrival());
-        carpoolFromDB.setEstimadedPrice(carpoolToUpdate.getEstimadedPrice());
+        carpoolFromDB.setEstimatedDeparture(carpoolToUpdate.getEstimatedDeparture());
+        carpoolFromDB.setEstimatedPrice(carpoolToUpdate.getEstimatedPrice());
         carpoolFromDB.setSeatsAvailable(carpoolToUpdate.getSeatsAvailable());
 
         return carpoolRepository.save(carpoolFromDB);
     }
 
     @Override
-    public List<Carpool> findAvailableCarpools() {
+    public List<CarpoolResponseDTO> findAvailableCarpools() {
         log.info("Buscando por caronas que possuem assentos disponíveis");
-        return carpoolRepository.findAvailableCarpools();
+        return carpoolRepository.findAvailableCarpools().stream()
+                .map(carpool -> CarpoolMapper.toCarpoolResponseDTO(carpool)).collect(Collectors.toList());
     }
 
     @Override
     @Transactional
     public Carpool saveCarpool(Carpool carpool) {
-        log.info("Salvando carona [{}]" , carpool.getId());
+        log.info("Salvando carona [{}]", carpool.getId());
         return carpoolRepository.save(carpool);
+    }
+
+    @Override
+    public Carpool updateCarpool(Long id, CarpoolUpdateRequestDTO carpoolUpdateRequestDTO) {
+        Carpool carpoolFromDB = findCarpoolById(id);
+
+        if (carpoolFromDB.getReservations().isEmpty()) {
+            throw new InvalidOperationException("Sua carona já possui reservas. Não é possivel atualiza-la");
+        }
+        carpoolFromDB.setEstimatedDeparture(carpoolUpdateRequestDTO.estimatedDeparture());
+        carpoolFromDB.setEstimatedPrice(carpoolUpdateRequestDTO.estimatedPrice());
+        carpoolFromDB.setSeatsAvailable(carpoolUpdateRequestDTO.seatsAvailable());
+
+        return carpoolRepository.save(carpoolFromDB);
+
     }
 }
